requirements/common.txt
# Flask
Flask==1.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.6.0
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-PageDown==0.2.2
Flask-HTTPAuth==3.2.3

# Search
jieba==0.39

# Markdown
Markdown==2.6.11
Pygments==2.2.0
bleach==2.1.3

# Email
sendgrid==5.3.0

# env
python-dotenv==0.8.2

# celery
redis==2.10.6
# hiredis
celery==4.1.0
